+++ CHANGELOG.md
@@ -6,7 +6,7 @@ If you like this project and find it useful, please consider giving it a star on
 
 ## [1.1.3] - 2024-12-15
 
-- [package]: Added thermostat Heat only.
+- [package]: Added thermostat Heat only with two external temperature sensors (tagged like Indoor and Outdoor)
 - [package]: Added thermostat Cool only.
 - [package]: Updated dependencies.